Retrieve Prometheus annotation from Method
The Prometheus annotation can only be applied to methods. However, the
Prometheus Dynamic Feature attempts to retrieve this annotation from the
Resource Class (on configuration). This means that the annotation will
never be successfully retrieved since it can never be applied on the
class level.

This commit changes the Prometheus Dynamic Feature to start retrieving
the annotation from the Resource Method instead of the Class.